Square In-app Payments example
