--- jacoby_arts/lib/pages/SquareCheckoutPage.dart
@@ -1,31 +1,0 @@
-import 'package:square_in_app_payments/in_app_payments.dart';
-import 'package:square_in_app_payments/models.dart';
-
-
-    Future<void> _onStartCardEntryFlow() async {
-    await InAppPayments.startCardEntryFlow(
-        onCardNonceRequestSuccess: _onCardEntryCardNonceRequestSuccess,
-        onCardEntryCancel: _onCancelCardEntryFlow);
-  }
-    void _onCancelCardEntryFlow() {
-    // Handle the cancel callback
-  }
-    void _onCardEntryCardNonceRequestSuccess(CardDetails result) async {
-    try {
-      // take payment with the card nonce details
-      // you can take a charge
-      // await chargeCard(result);
-
-      // payment finished successfully
-      // you must call this method to close card entry
-      void _onCardEntryComplete() {
-    // Update UI to notify user that the payment flow is finished successfully
-  }
-      InAppPayments.completeCardEntry(
-          onCardEntryComplete: _onCardEntryComplete);
-    } on Exception catch (ex) {
-      // payment failed to complete due to error
-      // notify card entry to show processing error
-      InAppPayments.showCardNonceProcessingError(ex.message);
-    }
-  }
